code/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
torchaudio>=0.9.0
Pillow>=8.0.0
matplotlib>=3.5.0
opencv-python>=4.5.0
numpy>=1.21.0
scipy>=1.7.0
tqdm>=4.60.0
lpips>=0.1.4
pytorch-fid>=0.3.0
scikit-image>=0.18.0
torchmetrics>=0.7.0
